+++ src/sentry_logger.c
@@ -56,11 +56,14 @@ sentry__logger_defaultlogger(
     size_t len = strlen(prefix) + strlen(priority)
         + sentry__guarded_strlen(message) + 2;
     char *format = sentry_malloc(len);
-    if (format) {
-        snprintf(format, len, "%s%s%s\n", prefix, priority, message);
-        vfprintf(stderr, format, args);
-        sentry_free(format);
+    if (!format) {
+        return;
     }
+    snprintf(format, len, "%s%s%s\n", prefix, priority, message);
+
+    vfprintf(stderr, format, args);
+
+    sentry_free(format);
 }
 
 #endif