+++ tests/test_integration_http.py
@@ -43,7 +43,7 @@ pytestmark = pytest.mark.skipif(not has_http, reason="tests need http")
 
 # fmt: off
 auth_header = (
-    "Sentry sentry_key=uiaeosnrtdy, sentry_version=7, sentry_client=sentry.native/0.8.5"
+    "Sentry sentry_key=uiaeosnrtdy, sentry_version=7, sentry_client=sentry.native/0.9.1"
 )
 # fmt: on
 
@@ -1272,7 +1272,7 @@ def test_capture_with_scope(cmake, httpserver):
     run(
         tmp_path,
         "sentry_example",
-        ["log", "attachment", "capture-with-scope"],
+        ["log", "attach-to-scope", "capture-with-scope"],
         check=True,
         env=dict(os.environ, SENTRY_DSN=make_dsn(httpserver)),
     )