+++ tests/unit/test_value.c
@@ -484,9 +484,11 @@ SENTRY_TEST(value_json_deeply_nested)
     }
 
     sentry_jsonwriter_t *jw = sentry__jsonwriter_new_sb(NULL);
+    TEST_ASSERT(!!jw);
     sentry__jsonwriter_write_value(jw, root);
     size_t serialized_len = 0;
     char *serialized = sentry__jsonwriter_into_string(jw, &serialized_len);
+    TEST_ASSERT(!!serialized);
     sentry_value_decref(root);
 
     sentry_value_t parsed = sentry__value_from_json(serialized, serialized_len);
@@ -683,7 +685,7 @@ SENTRY_TEST(value_get_by_null_key)
 
 SENTRY_TEST(value_set_stacktrace)
 {
-#if defined(SENTRY_PLATFORM_NX) || defined(SENTRY_PLATFORM_PS)
+#if defined(SENTRY_PLATFORM_NX)
     return SKIP_TEST();
 #endif
     sentry_value_t exc