+++ tests/win_utils.py
@@ -1,7 +1,7 @@
 import pathlib
 import win32api
 
-sentry_version = "0.8.5"
+sentry_version = "0.9.1"
 
 
 def check_binary_version(binary_path: pathlib.Path):